* master: (77 commits)
  Dimer::nclosed_ was redundant
  Dimer::nbasis_ and Dimer::nele_ were redundant
  Dimer::dimerbasis_ was redundant
  removing redundant coeffs_ and scoeff_ in Dimer class
  bug fixed.
  extract anonymous namespaces with doxygen
  fixing linear dependency issues. Fixes issue #11.
  Coeff should not have Geometry. MoldenIn had to be modified. In addition Coeff now has a move constructor.
  minor cleanup
  oops
  another round
  exterminating raw pointers step by step. Also implements get_real_part and get_imag_part in RDM.
  removing a few extraneous allocations
  Range::extent(int i) seems an alias to Range::range(int i).size() in my context
  missed one cosmetic change
  reorganizing MEH to have a base class so all of the MEH methods could be handled similarly
  implementing 3index*3index = 2*index in btas::contract
  contract for 3index * 2index = 3index
  BTAS updates
  Cleaning up the cherry pick - Numerical stability bug fix
  ...

+++ src/math/xyzfile.h
@@ -28,6 +28,7 @@
 #define __SRC_MATH_XYZFILE_H
 
 #include <src/math/matrix.h>
+#include <src/math/matop.h>
 
 namespace bagel {
 
@@ -41,7 +42,7 @@ class XYZFile : public Matrix {
 
     std::shared_ptr<XYZFile> clone() const;
 
-    void print(const std::string in = "", const size_t dum = 0) const override;
+    void print(const std::string in = "", const int dum = 0) const override;
 
     // this function assumes that double[] has data_.size()*data_size() elements.
     std::shared_ptr<XYZFile> transform(const std::shared_ptr<const Matrix>, const bool transpose) const;
